+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ## 0.3.2
 
-* BUG: Fix bug where `pytoil info` on a repo with no OSS license would cause an error. Not just displays `None` for `license`.
+* BUG: Fix bug where `pytoil info` on a repo with no OSS license would cause an error. Now just displays `None` for `license`.
 
 ## 0.3.1